requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
PyMySQL==1.1.1
python-dotenv==1.1.0
requests==2.32.3
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.3.0
